Move the hexctl test harness out from under its own bound
plugins/hexaemeron/tests/test_hexctl.py was 262,000 bytes against the
262,144-byte limit read_markdown enforces on a coverage evidence target. That
is 144 bytes, about two lines. Every guard added to the controller lands in
that file, so the next one fails the coverage command with a PM003 finding that
names the file's size rather than the change that crossed it.

The largest object in the file was the fixture the tests run on, not the tests:
HexctlCase at 32,229 bytes holding none of its own, over the 1,310-byte
OriginCheckoutMixin. Both move to hexctl_harness.py with the module-level
constants and helpers they close over. Nothing in the prelude referenced a name
defined below it, so this is a relocation rather than a rewrite -- the bytes
are identical on both sides of the two docstrings.

test_hexctl.py drops to 227,543 bytes, leaving 34,601 of headroom instead of
144, and re-exports the moved names, because test_fiat_skill,
test_hexctl_frontier_receipt and the case builders that take a namespace import
them from this module and the fixture is still, to them, where the tests are.

No coverage row moves. All fifteen rows citing this file name a test_ selector
and neither class carried one, so nothing had to be re-pointed. 376 tests still
collect, the count the file had before.

Two extractions have already come out of this file, into host_identity_cases.py
and replacement_object_cases.py, and each bought a few hundred bytes. This one
is the harness, so it buys two orders of magnitude more. It does not make the
bound elastic: the file will fill again, and a per-file bound on a module that
every controller guard lands in is the thing that recurs, not the byte count.

Refs #902.

Wildcat-Origin: shoggoth
